lookup table for hind foothold

+++ src/stair_main.cpp
@@ -153,7 +153,7 @@ int main(int argc, char** argv) {
                 }//end if
 
                 eta_list = walk_gait.step();
-                command_pitch_CoM << command_count << "," << trigger_msg.enable << "," << 0.0 << "," << exp_robot_x << "," << stand_height << "\n";
+                command_pitch_CoM << command_count << "," << (int)trigger_msg.enable << "," << 0.0 << "," << exp_robot_x << "," << stand_height << "\n";
                 command_count ++;
                 break;
             case STAIR:
@@ -168,7 +168,7 @@ int main(int argc, char** argv) {
                 eta_list = stair_climb.step();
                 pitch = stair_climb.get_pitch();
                 CoM = stair_climb.get_CoM();
-                command_pitch_CoM << command_count << "," << trigger_msg.enable << "," << pitch << "," << CoM[0] << "," << CoM[1] << "\n";
+                command_pitch_CoM << command_count << "," << (int)trigger_msg.enable << "," << pitch << "," << CoM[0] << "," << CoM[1] << "\n";
 
                 command_count ++;
                 break;
